+++ src/MillionaireGame/Forms/GameLogWindow.cs
@@ -1,5 +1,6 @@
 using System.Text;
 using MillionaireGame.Utilities;
+using MillionaireGame.Helpers;
 
 namespace MillionaireGame.Forms;
 
@@ -35,6 +36,10 @@ public partial class GameLogWindow : Form
 
         // Write header
         LogHeader();
+        
+        // Apply icon after everything is set up
+        IconHelper.ApplyToForm(this);
+        Log($"GameLogWindow constructor completed. Icon applied. Form.Icon is {(Icon != null ? "SET" : "NULL")}", LogLevel.DEBUG);
     }
 
     private void InitializeComponent()
@@ -48,17 +53,6 @@ public partial class GameLogWindow : Form
         
         FormBorderStyle = FormBorderStyle.Sizable;
         MinimumSize = new Size(400, 300);
-        
-        // Set icon
-        try
-        {
-            var iconPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "lib", "image", "logo.ico");
-            if (File.Exists(iconPath))
-            {
-                Icon = new Icon(iconPath);
-            }
-        }
-        catch { /* Ignore icon load errors */ }
         
         // Don't close, just hide
         FormClosing += (s, e) =>